--- README.md
@@ -1,2 +1,0 @@
-# fancy-todo
-Membuat website untuk mencatat hal - hal menarik untuk dilakukan
