+++ Server/package.json
@@ -22,13 +22,13 @@
   "dependencies": {
     "axios": "^0.21.1",
     "bcryptjs": "^2.4.3",
-    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "pg": "^8.5.1",
     "sequelize": "^6.3.5"
   },
   "devDependencies": {
+    "dotenv": "^8.2.0",
     "sequelize-cli": "^6.2.0"
   }
 }
